refactor: centralize database initialization
